requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-multiselectfield==0.1.12
idna==3.4
oauthlib==3.2.1
Pillow==9.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
python-dotenv==0.21.0
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23
urllib3==1.26.12
